BackEnd/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
python-multipart>=0.0.5
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests
python-jose[cryptography]
PyJWT
# Solo questa versione
git+https://github.com/openai/openai-agents-python.git 